requirements.txt
appdirs==1.4.3
bleach==2.0.0
dj-database-url==0.4.2
Django==1.11.1
git+https://github.com/eyesnz/django-address.git#egg=django-address
django-anymail==0.9
django-markdownify==0.1.0
enum34==1.1.6
geopy==1.11.0
html5lib==0.999999999
Markdown==2.6.8
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
requests==2.14.2
six==1.10.0
webencodings==0.5.1
whitenoise==3.3.0
psycopg2
gunicorn
mailchimp3
raven
python-cicero
djangorestframework
django-webpack-loader
django-rq
django-slack-oauth
django-redis-cache
drf-enum-field
django-dotenv
airtable-python-wrapper==0.11.0
social-auth-core
social-auth-app-django
django-taggit
django-taggit-serializer
